requirements.txt
flask
flask-login
flask-sqlalchemy
psycopg2-binary
cryptography
pyzotero
requests
feedgen
python-dotenv